third_party/blink/web_tests/fast/dom/shadow/hover-active-drag-distributed-nodes-expected.txt

CONSOLE WARNING: line 56: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
Top-level text Nested text
Other div
PASS backgroundColor is "rgb(0, 128, 0)"
PASS backgroundColor is "rgb(0, 255, 255)"
PASS backgroundColor is "rgb(0, 128, 0)"
PASS backgroundColor is "rgb(255, 0, 0)"
PASS backgroundColor is "rgb(0, 128, 0)"
PASS textColor is "rgb(0, 128, 0)"
PASS backgroundColor is "rgb(0, 255, 255)"
PASS textColor is "rgb(0, 255, 255)"
PASS backgroundColor is "rgb(0, 128, 0)"
PASS textColor is "rgb(0, 128, 0)"
PASS backgroundColor is "rgb(255, 0, 0)"
PASS textColor is "rgb(0, 0, 0)"
PASS backgroundColor is "rgb(0, 128, 0)"
PASS backgroundColor is "rgb(0, 255, 255)"
PASS backgroundColor is "rgb(0, 128, 0)"
PASS backgroundColor is "rgb(255, 0, 0)"
PASS successfullyParsed is true

TEST COMPLETE
